fin de los primeros

+++ src/ej04.py
@@ -1,3 +1,3 @@
-tempc = int(input('Ingrese temperatura en Celsius: '))
-tempf = int(tempc + 32)
+tempc = float(input('Ingrese temperatura en Celsius: '))
+tempf = tempc + 32
 print('Su temperatura en Fahrenheit es:', tempf)
